requirements.txt
matplotlib
pandas
plotly
python-dotenv
Requests
Sastrawi
streamlit
torch
transformers
tweepy
wordcloud
gdown
groq
